Update to pixman 0.10, with one small fix to the sse2 test in configure.ac.

+++ lib/pixman/pixman/pixman-image.c
@@ -28,7 +28,6 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "pixman.h"
 #include "pixman-private.h"
 
 static void
@@ -51,7 +50,7 @@ init_gradient (gradient_t     *gradient,
 	return FALSE;
 
     memcpy (gradient->stops, stops, n_stops * sizeof (pixman_gradient_stop_t));
-    
+
     gradient->n_stops = n_stops;
 
     gradient->stop_range = 0xffff;
@@ -75,7 +74,7 @@ static pixman_image_t *
 allocate_image (void)
 {
     pixman_image_t *image = malloc (sizeof (pixman_image_t));
-    
+
     if (image)
     {
 	image_common_t *common = &image->common;
@@ -134,7 +133,7 @@ pixman_image_unref (pixman_image_t *image)
 	if (image->type == BITS && image->bits.indexed)
 	    free (image->bits.indexed);
 #endif
-	
+
 #if 0
 	memset (image, 0xaa, sizeof (pixman_image_t));
 #endif
@@ -144,10 +143,10 @@ pixman_image_unref (pixman_image_t *image)
 		free (image->gradient.stops);
 	}
 
-	
+
 	if (image->type == BITS && image->bits.free_me)
 	    free (image->bits.free_me);
-	
+
 	free (image);
 
 	return TRUE;
@@ -163,9 +162,9 @@ pixman_image_create_solid_fill (pixman_color_t *color)
     pixman_image_t *img = allocate_image();
     if (!img)
 	return NULL;
-    
+
     init_source_image (&img->solid.common);
-    
+
     img->type = SOLID;
     img->solid.color = color_to_uint32 (color);
 
@@ -182,14 +181,14 @@ pixman_image_create_linear_gradient (pixman_point_fixed_t         *p1,
     linear_gradient_t *linear;
 
     return_val_if_fail (n_stops >= 2, NULL);
-    
+
     image = allocate_image();
-    
+
     if (!image)
 	return NULL;
 
     linear = &image->linear;
-    
+
     if (!init_gradient (&linear->common, stops, n_stops))
     {
 	free (image);
@@ -217,7 +216,7 @@ pixman_image_create_radial_gradient (pixman_point_fixed_t         *inner,
     radial_gradient_t *radial;
 
     return_val_if_fail (n_stops >= 2, NULL);
-    
+
     image = allocate_image();
 
     if (!image)
@@ -232,7 +231,7 @@ pixman_image_create_radial_gradient (pixman_point_fixed_t         *inner,
     }
 
     image->type = RADIAL;
-    
+
     radial->c1.x = inner->x;
     radial->c1.y = inner->y;
     radial->c1.radius = inner_radius;
@@ -245,7 +244,7 @@ pixman_image_create_radial_gradient (pixman_point_fixed_t         *inner,
     radial->A = (radial->cdx * radial->cdx
 		 + radial->cdy * radial->cdy
 		 - radial->dr  * radial->dr);
-    
+
     return image;
 }
 
@@ -262,7 +261,7 @@ pixman_image_create_conical_gradient (pixman_point_fixed_t *center,
 	return NULL;
 
     conical = &image->conical;
-    
+
     if (!init_gradient (&conical->common, stops, n_stops))
     {
 	free (image);
@@ -323,11 +322,11 @@ static void
 reset_clip_region (pixman_image_t *image)
 {
     pixman_region_fini (&image->common.clip_region);
-    
+
     if (image->type == BITS)
     {
 	pixman_region_init_rect (&image->common.clip_region, 0, 0,
-				 image->bits.width, image->bits.height);	
+				 image->bits.width, image->bits.height);
     }
     else
     {
@@ -345,10 +344,10 @@ pixman_image_create_bits (pixman_format_code_t  format,
     pixman_image_t *image;
     uint32_t *free_me = NULL;
 
-    /* must be a whole number of uint32_t's 
+    /* must be a whole number of uint32_t's
      */
     return_val_if_fail (bits == NULL ||
-			(rowstride_bytes % sizeof (uint32_t)) == 0, NULL); 
+			(rowstride_bytes % sizeof (uint32_t)) == 0, NULL);
 
     if (!bits && width && height)
     {
@@ -356,7 +355,7 @@ pixman_image_create_bits (pixman_format_code_t  format,
 	if (!bits)
 	    return NULL;
     }
-    
+
     image = allocate_image();
 
     if (!image) {
@@ -371,8 +370,8 @@ pixman_image_create_bits (pixman_format_code_t  format,
     image->bits.height = height;
     image->bits.bits = bits;
     image->bits.free_me = free_me;
-    
-    image->bits.rowstride = rowstride_bytes / sizeof (uint32_t); /* we store it in number
+
+    image->bits.rowstride = rowstride_bytes / (int) sizeof (uint32_t); /* we store it in number
 								  * of uint32_t's
 								  */
     image->bits.indexed = NULL;
@@ -398,7 +397,7 @@ pixman_image_set_clip_region (pixman_image_t    *image,
     else
     {
 	reset_clip_region (image);
-	
+
 	return TRUE;
     }
 }
@@ -423,7 +422,7 @@ pixman_image_set_transform (pixman_image_t           *image,
 	  { 0, 0, pixman_fixed_1 }
 	}
     };
-    
+
     image_common_t *common = (image_common_t *)image;
 
     if (common->transform == transform)
@@ -435,7 +434,7 @@ pixman_image_set_transform (pixman_image_t           *image,
 	common->transform = NULL;
 	return TRUE;
     }
-    
+
     if (common->transform == NULL)
 	common->transform = malloc (sizeof (pixman_transform_t));
     if (common->transform == NULL)
@@ -453,7 +452,7 @@ pixman_image_set_repeat (pixman_image_t  *image,
     image->common.repeat = repeat;
 }
 
-pixman_bool_t 
+pixman_bool_t
 pixman_image_set_filter (pixman_image_t       *image,
 			 pixman_filter_t       filter,
 			 const pixman_fixed_t *params,
@@ -477,7 +476,7 @@ pixman_image_set_filter (pixman_image_t       *image,
     }
 
     common->filter = filter;
-	
+
     if (common->filter_params)
 	free (common->filter_params);
 
@@ -518,7 +517,7 @@ pixman_image_set_alpha_map (pixman_image_t *image,
 			    int16_t         y)
 {
     image_common_t *common = (image_common_t *)image;
-    
+
     return_if_fail (!alpha_map || alpha_map->type == BITS);
 
     if (common->alpha_map != (bits_image_t *)alpha_map)
@@ -586,7 +585,7 @@ int
 pixman_image_get_stride (pixman_image_t *image)
 {
     if (image->type == BITS)
-	return image->bits.rowstride * sizeof (uint32_t);
+	return image->bits.rowstride * (int) sizeof (uint32_t);
 
     return 0;
 }
@@ -617,7 +616,7 @@ color_to_pixel (pixman_color_t *color,
     {
 	return FALSE;
     }
-    
+
     if (PIXMAN_FORMAT_TYPE (format) == PIXMAN_TYPE_ABGR)
     {
 	c = ((c & 0xff000000) >>  0) |
@@ -636,7 +635,7 @@ color_to_pixel (pixman_color_t *color,
     printf ("color: %x %x %x %x\n", color->alpha, color->red, color->green, color->blue);
     printf ("pixel: %x\n", c);
 #endif
-    
+
     *pixel = c;
     return TRUE;
 }
@@ -651,7 +650,7 @@ pixman_image_fill_rectangles (pixman_op_t		    op,
     pixman_image_t *solid;
     pixman_color_t c;
     int i;
-    
+
     if (color->alpha == 0xffff)
     {
 	if (op == PIXMAN_OP_OVER)
@@ -666,14 +665,14 @@ pixman_image_fill_rectangles (pixman_op_t		    op,
 	c.alpha = 0;
 
 	color = &c;
-	
+
 	op = PIXMAN_OP_SRC;
     }
 
     if (op == PIXMAN_OP_SRC)
     {
 	uint32_t pixel;
-	
+
 	if (color_to_pixel (color, &pixel, dest->bits.format))
 	{
 	    for (i = 0; i < n_rects; ++i)
@@ -681,7 +680,7 @@ pixman_image_fill_rectangles (pixman_op_t		    op,
 		pixman_region16_t fill_region;
 		int n_boxes, j;
 		pixman_box16_t *boxes;
-		
+
 		pixman_region_init_rect (&fill_region, rects[i].x, rects[i].y, rects[i].width, rects[i].height);
 		pixman_region_intersect (&fill_region, &fill_region, &dest->common.clip_region);
 
@@ -699,7 +698,7 @@ pixman_image_fill_rectangles (pixman_op_t		    op,
 	    return TRUE;
 	}
     }
-    
+
     solid = pixman_image_create_solid_fill (color);
     if (!solid)
 	return FALSE;
@@ -707,13 +706,13 @@ pixman_image_fill_rectangles (pixman_op_t		    op,
     for (i = 0; i < n_rects; ++i)
     {
 	const pixman_rectangle16_t *rect = &(rects[i]);
-	
+
 	pixman_image_composite (op, solid, NULL, dest,
 				0, 0, 0, 0,
 				rect->x, rect->y,
 				rect->width, rect->height);
     }
-    
+
     pixman_image_unref (solid);
 
     return TRUE;
